fixes with time and radians
