+++ sopp/custom_dataclasses/position.py
@@ -1,9 +1,91 @@
 from dataclasses import dataclass
 from typing import Optional
 
-import math
-from sopp.event_finder.event_finder_rhodesmill.support.satellite_link_budget_angle_calc import CartesianCoordinate
+import numpy as np
+from typing import List
 
+import math
+
+
+class CartesianCoordinate:
+    def __init__(self, x: float, y: float, z: float):
+        self.x = x  # x-coordinate
+        self.y = y  # y-coordinate
+        self.z = z  # z-coordinate
+
+    def set_coordinates(self, x, y, z):
+        self.x = x
+        self.y = y
+        self.z = z
+
+    def get_coordinates(self):
+        return self.x, self.y, self.z
+    
+    def pass_to_rotation_matrix(self, gamma: float, phi: float): #gamma and phi are in degrees when we input them, so we must convert to radians
+        gamma = np.deg2rad(gamma)
+        phi = np.deg2rad(phi)
+        matrixOne = np.array([
+            [np.cos(gamma), -np.sin(gamma), 0],
+            [np.sin(gamma), np.cos(gamma), 0],
+            [0, 0, 1]
+        ])
+
+        matrixTwo = np.array([
+            [np.cos(phi), 0, np.sin(phi)],
+            [0, 1, 0],
+            [-np.sin(phi), 0, np.cos(phi)]
+        ])
+        origMatrix = np.matmul(matrixOne, matrixTwo)
+        totMatrix = origMatrix.T
+        cartesianArray = np.array([self.x,self.y,self.z])
+        return self.apply_rotation(totMatrix,cartesianArray)
+    
+    def pass_to_gsrc_local_matrix(self, matrixOne: np.array, matrixTwo: np.array):
+        origMatrix = np.matmul(matrixOne,matrixTwo)
+        #totMatrix = origMatrix.T
+        cartesianArray = np.array([self.x,self.y,self.z])
+        return self.apply_rotation(origMatrix,cartesianArray)
+    
+    def apply_rotation(matrix: np.ndarray, vector: np.ndarray) -> np.ndarray:
+        """
+        Apply rotation matrix to a vector of Cartesian coordinates.
+    
+        Args:
+        - matrix (np.ndarray): 3x3 rotation matrix
+        - vector (np.ndarray): 1x3 vector of Cartesian coordinates [x, y, z]
+    
+        Returns:
+        - np.ndarray: Resulting vector after rotation
+        """
+        return np.matmul(matrix, vector)
+    
+    def cartesian_to_spherical(self) -> List[float]:
+        """
+        Convert Cartesian coordinates to spherical coordinates.
+        """
+
+        x = self.x
+        y = self.y
+        z = self.z
+
+        r = np.sqrt(x**2 + y**2 + z**2)  # Radial distance
+
+        theta = np.arccos(z/r)
+
+        if x > 0:
+            phi = np.arctan(y/x)
+        if x < 0 and y >= 0:
+            phi = np.arctan(y/x) + np.pi
+        if x < 0 and y < 0:
+            phi = np.arctan(y/x) - np.pi
+        if x == 0 and y > 0:
+            phi = np.pi/2
+        if x == 0 and y < 0:
+            phi = -np.pi/2
+        if x == 0 and y == 0:
+            phi = 0 # default value because this is undefined, maybe change this to some other value?
+
+        return [theta, phi]
 
 @dataclass
 class Position:
@@ -39,3 +121,4 @@ class Position:
         z = self.distance_km * math.cos(theta)
         
         return CartesianCoordinate(x, y, z)
+
